Ajuste no componente TableCell:  > Separado a lógica de formatação de valor do TableCell para componente TableCellValue, para possibilitar sua reutilização em componentes customizados de célula.  > Movido o array de formatações válidas de um TableCell para o Enum TableCellValueFormats.
