+++ src/js/enums/index.js
@@ -1,1 +1,2 @@
 export * from './input';
+export * from './table';
