+++ src/js/enums/table/index.js
@@ -1,0 +1,1 @@
+export { default as TableCellValueFormats } from './table_cell_value_formats';
