+++ src/js/enums/table/table_cell_value_formats.js
@@ -1,0 +1,10 @@
+export default [
+  'text',
+  'currency',
+  'number',
+  'percentage',
+  'boolean',
+  'date',
+  'datetime',
+  'time',
+];
